user-management. gitlab4j added

+++ src/main/java/com/redhat/labs/lodestar/participants/model/Engagement.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 public class Engagement {
 
     private String uuid;
-    private long projectId;
+    private int projectId;
     
 }
